Change: Scripts/Entry> change type of 'permission' to int

+++ Assets/Scripts/Entry.cs
@@ -9,7 +9,7 @@ public class Entry : MonoBehaviour
     public string path;
     public string name;
     public Entry  parent;
-    public byte permission;
+    public int permission;
     public System.DateTime lastTouched;
     public GameObject model;
